pyproject.toml
[tool.poetry]
name = "welsh"
version = "0.1.0"
description = ""
authors = ["Rich Wareham <rich@richwareham.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
pyyaml = "^6.0.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
